requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
